:art: refactor : Path 객체를 ShortestPathFinder 이름으로 개선
