+++ http/path.http
@@ -1,7 +1,7 @@
 ###
 # @name getShortestPath
-# @description Retrieve shortest path
-GET http://localhost:8080/path
+# @description Retrieve shortest shortestPathFinder
+GET http://localhost:8080/shortestPathFinder
 Content-Type: application/json
 Accept: application/json
